@NHDaly 2025-08-26
> We should probably be enforcing that any options (like precompile dispatch counting) are forcibly disabled while running the compiler. You do not want to ignore these after the fact or mark them differently in the output, since then you will miss actual real dispatches that occur at runtime simply because they also could occur during compile.

Ah, that's an interesting point. So you're saying we should somehow "forget" that we've compiled this, and then the first time it's dispatched to again in the future -- outside the compiler -- we should log it as if we'd precompiled it then? 🤔 that seems difficult. I do see the merit, but that seems like a much harder code change, and i'm not _entirely_ sure of the merit. 

I think that probably people shouldn't skip those during compilation _replay_ for snooping, to be careful about the situation you described. We mostly just wanted to annotate these to avoid double-counting during post-processing aggregation for our observability data: "how much time did this engine spend compiling, according to the precompile traces?" (We also look at other metrics like `Base.cumulative_compile_time_ns()`, but it's helpful to correlate these.)

Maybe this needs a paragraph in a documentation page about the --trace-compile feature to explain the nuances above?