@NHDaly 2025-09-06
So, @IanButterworth / @gbaraldi: Any thoughts on this? I agree with @vtjnash that long-term we might want to think about not emitting precompile statements at all for these, or doing some other kind of math to subtract the compilation time from their parents or something like that. But for now, i think just tagging these so that the user can differentiate them is a good improvement. Thoughts from your end?