@IanButterworth 2025-09-06
Perhaps adding it to the docs would help outlining the motivation to highlight it? Without docs I can see users like myself not understanding what it means